+++ src/Components/AccountDropdownMenu.js
@@ -1,10 +1,9 @@
 import React from "react";
 import { connect } from "react-redux";
 
-import { AccountDropdown, GhostButton, MenuIcon } from "../Styles";
+import { AccountDropdown, GhostButton, MenuIcon, Help } from "../Styles";
 import DashAmount from "./DashAmount";
 import IdentityCredit from "./IdentityCredit";
-import HelpInfo from "./HelpInfo";
 
 function AccountDropdownMenu({
   isVisible,
@@ -21,14 +20,7 @@ function AccountDropdownMenu({
     <AccountDropdown isVisible={isVisible} onClick={hide}>
       <h5>Dash Balance</h5>
       <DashAmount unconfirmed={unconfirmed}>{confirmed}</DashAmount>
-      <h5>
-        Platform Credits
-        <HelpInfo>
-          Platform Credits allow you to interract with the Dash Platform.
-          Anytime you interract, you will spend a small amount of Platform
-          Credits. No credits are used when just viewing content.
-        </HelpInfo>
-      </h5>
+      <h5>Platform Credits</h5>
       <IdentityCredit>{identityBalance}</IdentityCredit>
       {/*<ul>
         <li>